os/fs/FS.h: fix do-while loop
In case do_submit from libaio fails and returns -EAGAIN the
called 'continue' breaks completely out of the loop since the
loop statement is always false. In this case there is no second
attempt as the code tries. Set the loop condition to true to
fix it.

Fix for:

CID 1322822 (#1 of 1): Structurally dead code (UNREACHABLE)
 continue_in_do_while_false: A continue statement within a do-while loop
 only continues execution of the loop body if the loop continuation
 condition is still true. Since the condition will never be true in a
 "do ... while (false);" loop the continue statement has the same effect
 as a break statement. Did you intend execution to continue at the top of
the loop?

Signed-off-by: Danny Al-Gaaf <danny.al-gaaf@bisect.de>